DocScanner/requirements.txt
numpy==1.19.0
opencv_python==4.2.0.34
Pillow==9.4.0
scikit_image==0.17.2
skimage==0.0
torch==1.5.1+cu101
